Address Issues identified by SpotBugs
